Handling of cooments from github
