+++ features/steps/step_checking-user-name-on-people-page.py
@@ -1,10 +1,10 @@
 from behave import when, then
-import features.steps.step_checking_status as cs
+# import features.steps.step_checking_status as cs
 
 
 @when("we compare name of current user with name on Welcome title")
 def step_impl(context):
-    assert cs.FULL_NAME in context.authorized_page.get_page_head()
+    assert context.hipchat_full_name in context.authorized_page.get_page_head()
 
 
 @when("we move to user page (on People Page)")
@@ -14,4 +14,4 @@ def step_impl(context):
 
 @then("we compare name of current user with name in user page")
 def step_impl(context):
-    assert cs.FULL_NAME in context.people_page.label_page_head_on_user_page()
+    assert context.hipchat_full_name in context.people_page.label_page_head_on_user_page()
