+++ pages/lobby_page.py
@@ -14,10 +14,32 @@ class LobbyPage(Page):
     """
 
     url = '/chat/lobby'
-    FULL_NAME = None
+    # FULL_NAME = None
     status_shortcuts = {'available': 'icon-avail',
                         'away': 'icon-xa',
                         'do not disturb': 'icon-dnd'}
+
+    class lobby_icon_changed(object):
+        def __init__(self, lobby_page, div, icon_status):
+            self.div = div
+            self.icon_status = icon_status
+            self.lobby_page = lobby_page
+
+        def __call__(self, *args, **kwargs):
+            element = self.lobby_page.find_ico_in_div(self.div)
+            icon_str = str(element.get_attribute('xlink:href'))
+            if self.icon_status not in icon_str:
+                return icon_str
+            else:
+                return False
+
+    # def lobby_icon_changed(self, div, icon_status):
+    #     element = self.find_ico_in_div(div)
+    #     icon_str = str(element.get_attribute('xlink:href'))
+    #     if icon_status not in icon_str:
+    #         return icon_str
+    #     else:
+    #         return False
 
     def open_rooms_list(self):
         self.context.wait.until(lambda driver: driver.find_element_by_id('status_dropdown'))
