+++ pages/people_page.py
@@ -2,7 +2,7 @@ from selenium.webdriver.common.keys import Keys
 from .base_page import Page
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support import expected_conditions as EC
-import features.steps.step_checking_status as cs
+# import features.steps.step_checking_status as cs
 import random
 import string
 
@@ -32,7 +32,7 @@ class PeoplePage(Page):
         self.context.wait.until(EC.visibility_of_element_located((By.XPATH, '//ol[@class="aui-nav-pagination"]')))
 
     def open_user_page(self):
-        self.context.driver.find_element_by_xpath('//a[contains(text(), "' + cs.FULL_NAME + '")]').click()
+        self.context.driver.find_element_by_xpath('//a[contains(text(), "' + self.context.hipchat_full_name + '")]').click()
 
     def label_page_head_on_user_page(self):
         return self.context.driver.find_element_by_css_selector("div.aui-item > h2").text
